feat: add mac bar to terminal component
